+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Brackets, extended edition".
 - Tests for "Target firing".
 - Tests for "Chained matrix products".
+- Tests for "Fun with set theory".
 
 ## [1.12.0] - 2022-09-01
 ### Added